+++ src/rules/selector-no-type/__tests__/index.js
@@ -50,6 +50,9 @@ testRule(rule, {
   }, {
     code: "@keyframes spin { to {} from {} }",
   }, {
+    code: "@include keyframes(identifier) { to, 50.0% {} 50.01% {} 100% {} }",
+    description: "non-standard usage of keyframe selectors",
+  }, {
     code: ":root { --custom-property-set: {} }",
   } ],