+++ src/rules/selector-no-type/index.js
@@ -1,6 +1,6 @@
 import { get } from "lodash"
 import {
-  isKeyframeRule,
+  isKeyframeSelector,
   isStandardSyntaxRule,
   isStandardSyntaxSelector,
   isStandardSyntaxTypeSelector,
@@ -33,10 +33,11 @@ export default function (on, options) {
 
     root.walkRules(rule => {
 
+      const { selector, selectors } = rule
+
       if (!isStandardSyntaxRule(rule)) { return }
-      if (isKeyframeRule(rule)) { return }
-      const { selector } = rule
       if (!isStandardSyntaxSelector(selector)) { return }
+      if (selectors.some(s => isKeyframeSelector(s))) { return }
 
       parseSelector(selector, result, rule, selectorAST => {
         selectorAST.walkTags(tag => {