+++ src/rules/selector-type-no-unknown/index.js
@@ -56,9 +56,11 @@ export default function (actual, options) {
     if (!validOptions) { return }
 
     root.walkRules(rule => {
+      const { selector, selectors } = rule
+
       if (!isStandardSyntaxRule(rule)) { return }
-      const { selector } = rule
       if (!isStandardSyntaxSelector(selector)) { return }
+      if (selectors.some(s => isKeyframeSelector(s))) { return }
 
       parseSelector(selector, result, rule, selectorTree => {
         selectorTree.walkTags(tagNode => {
@@ -66,7 +68,6 @@ export default function (actual, options) {
 
           const tagName = tagNode.value
           const tagNameLowerCase = tagName.toLowerCase()
-          if (isKeyframeSelector(tagName)) { return }
 
           if (htmlTags.indexOf(tagNameLowerCase) !== -1
             || svgTags.indexOf(tagNameLowerCase) !== -1